+++ js/popup.js
@@ -302,7 +302,8 @@ $(document).ready(function() {
             chrome.tabs.sendMessage(tabs[0].id, { type: 'autoEnRoute' });
         });
     });
-
+    
+    // manually print a label
     $('#sendPrintLabel').click(function() {
         chrome.tabs.query({active: true, currentWindow: true}, function(tabs) {
             chrome.tabs.sendMessage(tabs[0].id, { type: 'printLabel' + $('#printLabel').val() });